redis_kafka_mongo/requirements.txt
tqdm
click
redis
aiohttp
pymongo
kafka-python